requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
pillow
pytesseract
httpx